% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{add_response_columns}
\alias{add_response_columns}
\title{Compute the corresponding observed response values for a set of signals.}
\usage{
add_response_columns(
  signals,
  response_name,
  forecast_date,
  incidence_period,
  ahead
)
}
\arguments{
\item{signals}{Wide data frame of signals.  We expect the following columns:
\itemize{
\item{\code{geo_value}}{Strings of geographic locations.}
\item{\code{time_value}}{Dates of training data.}
\item{Covariate columns}{Columns with names of the form \code{value-{days}:{signal}} or
\verb{value+0:\{signal\} whose values correspond to }{signal}\verb{}{days}\code{before}time_value`}
}}

\item{response_name}{Name of response signal.  The value of the response variable on a single
day should be located in column \code{value+0:{response_name}} of \code{signals}.}

\item{forecast_date}{Date on which the forecast will be made}

\item{incidence_period}{Time period over which the response should be summed.}

\item{ahead}{integer or vector of integer ahead values}
}
\value{
Data frame of signals equal to \code{signals} with \code{length(ahead)} additional columns with
names \code{response+{a}:{response_name}} corresponding to the response variable at ahead \code{a}.
}
\description{
The response for an observed data point at time \code{t} is defined as the mean of observed values on
days \code{t + s} through \code{t + e} (inclusive) multiplied by \code{e - s}, where \code{s} and \code{e} are the number
of days from \code{forecast_date} to the start and end, respectively, of the next full
\code{incidence_period}.  We choose to take the mean and multiply by the incidence period length
rather than simply sum the observed values in order to avoid treating missing values as 0.
}
