% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-object-modifiers.R
\name{add_scores}
\alias{add_scores}
\title{Calculates and adds the scores to a fitted model object}
\usage{
add_scores(object, func, x = coef(object), log_transform = is_bgw(object), ...)
}
\arguments{
\item{object}{A fitted model object}

\item{func}{A function with real (vector) results returning the function
values at the observation level. This is either the log-likelihood function,
or, as expected by \code{\link[bgw:bgw_mle]{bgw::bgw_mle()}}, the likelihood function (probabilities).}

\item{x}{A real or real vector argument to func, indicating the point at
which the gradient is to be calculated. Defaults to the estimated
coefficients.}

\item{log_transform}{A logical value indicating if \code{func} returns
likelihoods that must be log transformed before differentiating. Defaults
to \code{TRUE} for objects of class \code{bgw_mle} and \code{FALSE} otherwise.}

\item{...}{Additional arguments passed to \code{\link[numDeriv:jacobian]{numDeriv::jacobian()}}.}
}
\value{
A fitted model object with the scores added to the
object.
}
\description{
Calculates and adds the scores to a fitted model object.
The scores are the gradient observations/first derivatives of the
observation-level log-likelihood function. The function is a wrapper around
\code{\link[numDeriv:jacobian]{numDeriv::jacobian()}}.
}
