% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-generics.R
\name{AIC.bgw_mle}
\alias{AIC.bgw_mle}
\title{AIC of a \code{bgw_mle} object}
\usage{
\method{AIC}{bgw_mle}(object, ..., k = 2)
}
\arguments{
\item{object}{An object of class \code{bgw_mle}}

\item{...}{Additional arguments}

\item{k}{The penalty parameter used. The default value is k = 2 for the
classical AIC}
}
\value{
A single double
}
\description{
Calculate the AIC of the model.
}
