% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-generics.R
\name{BIC.bgw_mle}
\alias{BIC.bgw_mle}
\title{BIC of a \code{bgw_mle} object}
\usage{
\method{BIC}{bgw_mle}(object, ...)
}
\arguments{
\item{object}{An object of class \code{bgw_mle}}

\item{...}{Additional arguments}
}
\description{
Calculate the BIC of the model.
}
