% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich-estimator.R
\name{bread}
\alias{bread}
\title{Bread for sandwiches}
\usage{
bread(x, n, ...)
}
\arguments{
\item{x}{A fitted model object}

\item{n}{The number of (clustered) observations}

\item{...}{Additional arguments passed to methods}
}
\value{
A matrix containing the expectation of the negative derivative of
the estimatino function. The default is to return the
}
\description{
An implementation of the bread for sandwiches function to use when calculating
the robust variance-covariance matrix for the sandwich estimator when using
the BGW algorithm
}
