% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmtools-package.R
\docType{package}
\name{cmtools-package}
\alias{cmtools}
\alias{cmtools-package}
\title{cmtools: Choice modeling and visualization in R}
\description{
A collection of tools for choice modeling and visualization of choice data in R using syntax consistent with the tidyverse.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://edsandorf.me/cmtools/}
}

}
\author{
\strong{Maintainer}: Erlend Dancke Sandorf \email{erlend.dancke.sandorf@nmbu.no} (\href{https://orcid.org/0000-0001-7521-4773}{ORCID})

}
\keyword{internal}
