% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-generics.R
\name{coef.bgw_mle}
\alias{coef.bgw_mle}
\title{Coefficients of a \code{bgw_mle} object}
\usage{
\method{coef}{bgw_mle}(object, ...)
}
\arguments{
\item{object}{An object of class \code{bgw_mle}}

\item{...}{Additional arguments}
}
\value{
a numeric vector
}
\description{
Retrieve the estimated coefficients of the model
}
