% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{confintVector}
\alias{confintVector}
\title{Confidence interval of a vector}
\usage{
confintVector(x, conf.level = 0.95)
}
\arguments{
\item{x}{A numeric vector}

\item{conf.level}{confidence level of the vector}
}
\value{
A vector of the lower/upper limits of the confidence interval
}
\description{
Calculates confidence interval of a vector using a normal distribution
}
