% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{converged}
\alias{converged}
\alias{converged.bgw_mle}
\alias{converged.maxLik}
\alias{converged.default}
\title{Check if model has converged}
\usage{
converged(x, ...)

\method{converged}{bgw_mle}(x, ...)

\method{converged}{maxLik}(x, ...)

\method{converged}{default}(x, ...)
}
\arguments{
\item{x}{A model object}

\item{...}{Additional arguments}
}
\value{
A logical value indicating if the model has converged
}
\description{
Using the return code from the optimizer, check if the code indicates that
the optimization converged to a local optimum.
}
\details{
Objects of class \code{bgw_mle} are considered converged for the following codes:
3 - X-convergence
4 - Relative function convergence
5 - X- and relative function convergence
6 - Absolute function convergence

Objects of class \code{maxLik} are considered converged for the following codes:
0 - Successful convergence (optim based methods, e.g. BFGS)
1 - Gradient close to zero
2 - Successive function values within tolerance limit
8 - Successive function values within relative tolerance limit
}
