% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{convertFilepath}
\alias{convertFilepath}
\title{Convert Windows Filepath to Character String for R}
\usage{
convertFilepath()
}
\value{
A converted character string with backslashes converted to forward slash.
}
\description{
To help copy/paste a filepath in Windows containing backslashes ('\')
that can be passed to R functions.
}
