% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition.R
\name{convertProbRR}
\alias{convertProbRR}
\title{Convert Probability with Relative Risk}
\usage{
convertProbRR(rr, pC, t)
}
\arguments{
\item{rr}{Relative risk}

\item{pC}{Input probability}

\item{t}{Time frame of relative risk}
}
\value{
A converted probability in the time frame of the input probability
}
\description{
Converts a probability using a relative risk from a different time frame
}
