% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covidhub_probs}
\alias{covidhub_probs}
\title{Probabilities (quantile levels) for
\href{https://github.com/reichlab/covid19-forecast-hub}{COVID-19 Forecast Hub}
forecasts}
\format{
Numeric vector with 23 entries.
}
\usage{
covidhub_probs
}
\description{
Probabilities (quantile levels) for forecasts in COVID Hub format: 0.01,
0.025, then evenly spaced from 0.05 to 0.95 in increments of 0.05, then
0.975, 0.99.
}
\keyword{datasets}
