% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{create_train_and_predict_matrices}
\alias{create_train_and_predict_matrices}
\title{Create training and testing data matrices and a training response vector for a given ahead.}
\usage{
create_train_and_predict_matrices(lagged_df, ahead, training_window_size)
}
\arguments{
\item{lagged_df}{Data frame of lagged data.  It should have the following columns:
\itemize{
\item{\code{geo_value}}{Strings of geographic locations.}
\item{\code{time_value}}{Dates of training data.}
\item{Covariate columns}{Columns with names of the form \code{value-{days}:{signal}} or
\verb{value+0:\{signal\} whose values correspond to }{signal}\verb{}{days}\code{before}time_value\verb{\} \\item\{Response columns\}\{Columns with names of the form }response+{n}:{response}\verb{whose values correspond to}{response}\verb{}{n}\verb{incidence period units after}time_value\verb{.\} \} A data frame in this format can be made using }covidcast::aggregate_signals()\code{and}modeltools::get_response_columns()`.}

\item{ahead}{Number of incidence period units (i.e., epiweeks, days, etc.) ahead to forecast}

\item{training_window_size}{Size of the local training window in days to use. For example, if
\code{training_window_size = 14}, then to make a 1-day-ahead forecast on December 15, we train on
data from December 1 to December 14.}
}
\value{
Named list with entries:
\itemize{
\item{\code{train_x}}{Matrix of training data whose columns correspond to the
\code{value-{days}:{signal}} columns in \code{lagged_df}.  The training data consists of the
latest date with an non-null response, plus all data from the \code{training_window_size}
days prior to it.}
\item{\code{train_y}}{Vector of response data from the \code{response+{ahead}:{response}} column of
\code{lagged_df} corresponding to the rows of \code{train_x}.}
\item{\code{predict_x}}{Matrix of prediction data in the same format as \code{train_x}.  The
prediction data contains the most recent \code{training_window_size} days.}
\item{\code{predict_geo_values}}{Vector of \code{geo_values} corresponding to the rows of \code{predict_x}.}
\item{\code{train_end_date}}{latest \code{time_value} used in the training period}
}
}
\description{
Create training and testing data matrices and a training response vector for a given ahead.
}
\examples{
\dontrun{
create_train_and_predict_matrices(
  tibble(
    geo_value = rep(c("az", "wv"), 5),
    time_value = rep(
      as.Date(c("2021-01-25", "2021-01-26", "2021-01-27", "2021-01-28", "2021-01-29")),
      each = 2),
    `value-2:signal_1` = seq(-3, 6),
    `value-1:signal_1` = seq(-1, 8),
    `value+0:signal_1` = seq(1, 10),
    `response+2:signal_1` = c(seq(5, 10), rep(NA, 4))
  ),
  ahead = 2,
  training_window_size = 1)
)
}

}
