% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{create_train_matrices}
\alias{create_train_matrices}
\title{Create training data matrix and a training response for given aheads.}
\usage{
create_train_matrices(lagged_df, ahead, n_days, train_end_date)
}
\arguments{
\item{lagged_df}{Data frame of lagged data as in \code{create_train_and_predict_matrices()}.}

\item{ahead}{Number of incidence period units (i.e., epiweeks, days, etc.)
ahead to forecast. Can be a single positive integer or a vector of
positive integers. Note that for each \code{{a}} in \code{ahead}, the column
\code{response+{a}:{response}} should be present in \code{lagged_df}.}

\item{n_days}{Number of days worth of data to pull.}

\item{train_end_date}{The last date to be included in the training data.}
}
\value{
A named list with entries:
\itemize{
\item \code{train_x}: Matrix of training data whose columns correspond to the
\code{value-{days}:{signal}} columns in \code{lagged_df}.  The training data consists of the
latest date with an non-null response, plus all data from the \code{training_window_size}
days prior to it.
\item \code{train_y}: Vector of response data from the \code{response+{ahead}:{response}} column of
\code{lagged_df} corresponding to the rows of \code{train_x}. If multiple ahead
values are provided, then this is a matrix instead.
\item \code{train_geo_values}: Vector of geo values corresponding to the rows
of \code{train_x}.
\item \code{train_time_values}: Vector of time values corresponding to the rows
of \code{train_x}.
\item \code{train_end_date}: Latest \code{time_value} used in the training period.
}
}
\description{
Create training and data matrix and training response for a set of
given aheads. Works for both single ahead values and a vector of ahead
values. However, note that this function works different from
\code{create_train_and_predict_matrices()} for multiple ahead values. If
multiple ahead values are supplied, we return a matrix of responses
containing all aheads at once.
}
