% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Print package startup message}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{The library name}

\item{pkgname}{The package name}
}
\value{
Nothing
}
\description{
The function is called when the package is loaded through the \code{library} or
\code{require} functions. It prints a message to the console.
}
