% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estdistparam.R
\name{estBetaParam}
\alias{estBetaParam}
\title{Estimate Beta Distribution Parameters}
\usage{
estBetaParam(mu, sigma)
}
\arguments{
\item{mu}{Mean of the beta distribution. 
\emph{mu} must be between 0 and 1.}

\item{sigma}{Standard deviation of the beta distribution.
\emph{sigma} cannot have a negative value.}
}
\description{
Estimate parameters of a beta distribution using method of moments.
}
\details{
Beta(\eqn{\alpha},\eqn{\beta}) distribution has two shape parameters, \eqn{\alpha} and \eqn{\beta}, estimated as:
\deqn{\alpha = [(1-mu)/sigma^2 - 1/mu] * mu^2}
\deqn{\beta = alpha * (1/mu - 1)}
}
\references{
https://stats.stackexchange.com/questions/12232/calculating-the-parameters-of-a-beta-distribution-using-the-mean-and-variance
}
