% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estdistparam.R
\name{estGammaParam}
\alias{estGammaParam}
\title{Estimate Gamma Distribution Parameters}
\usage{
estGammaParam(mu, sigma)
}
\arguments{
\item{mu}{Mean of the gamma distribution. 
\emph{mu} cannot have a negative value.}

\item{sigma}{Standard deviation of the gamma distribution.
\emph{sigma} cannot have a negative value.}
}
\description{
Estimate parameters of a gamma distribution using method of moments.
}
\details{
\eqn{\alpha} (shape) and \eqn{\beta} (scale) parameters are estimated as:
\deqn{\alpha = (mu/sigma)^2}
\deqn{\beta = (sigma^2)/mu}
}
\references{
https://www.rocscience.com/help/swedge/webhelp/swedge/Gamma_Distribution.htm
}
