% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv.R
\name{estimate_deriv}
\alias{estimate_deriv}
\title{Estimate derivatives of values in `covidcast_signal` data frame}
\usage{
estimate_deriv(
  df,
  method = c("linear-reg", "smooth-spline"),
  n = 14,
  new_col = "deriv",
  deriv = 1,
  params = list()
)
}
\arguments{
\item{df}{The `covidcast_signal` data frame under consideration.}

\item{method}{One of "linear-reg" or "smooth-spline", indicating the method
to use for the derivative calculation. To estimate the derivative at any
time point, we run the given method on the last `n` days of data, and use
the corresponding predicted derivative (that is, the derivative of the
underlying estimated function, either linear or cubic spline) at the
current time point.}

\item{n}{Size of the local window (in days) to use. For example, if `n = 5`,
then to estimate the derivative on November 5, we train the given method on
data in between November 1 and November 5. Default is 14.}

\item{new_col}{String indicating the name of the new column that will contain
the derivative values. Default is "deriv"; note that we can set `new_col = 
"value"` to overwrite the existing "value" column.}

\item{deriv}{Order of derivative to estimate. For `method = "linear-reg"`,
the order can only be 1. For `method = "smooth-spline"`, the order can be
1 or 2.}

\item{params}{List of additional arguments to pass to `lsfit()` when `method
= "linear-reg"`, or `smooth.spline()` when `method = "smooth-spline"`.}
}
\value{
A data frame given by appending a new column to `df` named according
  to the `new_col` argument, containing the estimated derivative values.
}
\description{
Estimates derivatives of the values in a `covidcast_signal` data frame, using
a local (in time) linear regression or smoothing spline. (When multiple issue
dates are present, only the latest issue is considered.)  See the
[derivatives vignette]() for examples.
}
