% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estdistparam.R
\name{estLogNormParam}
\alias{estLogNormParam}
\title{Estimate Lognormal Distribution}
\usage{
estLogNormParam(mu, sigma = NULL, lower = NULL, upper = NULL)
}
\arguments{
\item{mu}{Mean of estimate.}

\item{sigma}{Standard deviation of estimate.}

\item{lower}{Lower limit of 95\% confidence interval of estimate.}

\item{upper}{Upper limit of 95\% confidence interval of estimate.}
}
\value{
Log-mean and Log-sd of the lognormal distribution.
}
\description{
Estimate parameters of a lognormal distribution using method of moments.
}
\details{
Parameters for a 95\% confidence interval are optional if mu and sigma are provided. 
If neither sigma nor 95\% confidence interval are provided, the function returns \emph{NULL}.

Where \emph{mu} and \emph{sigma} are provided, parameters are estimated by:
\deqn{log(\mu) = log(mu) - log(\sqrt{1 + sigma^2/mu^2})}
\deqn{log(\sigma) = \sqrt{log(1 + sigma^2/mu^2)}}

Where \emph{mu} and a 95\% confidence interval are provided, parameters are estimated by:
\deqn{log(\mu) = log(mu)}
\deqn{log(\sigma) = [(log(upper) - log(lower))/(2*1.959964)]}
}
\references{
https://en.wikipedia.org/wiki/Log-normal_distribution
}
