% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_forecaster.R
\name{example_forecaster}
\alias{example_forecaster}
\title{Simple example forecaster that forecasts confirmed cases.}
\usage{
example_forecaster(df_list, forecast_date)
}
\arguments{
\item{df_list}{List of downloaded covidcast signals}

\item{forecast_date}{Date from which the forecast should be made.

preds <- evalcast::get_predictions(modeltools::example_forecaster,
"example",
tibble(
data_source = "jhu-csse",
signal = "confirmed_incidence_num",
geo_type = "state",
start_date = as.Date("2021-01-01"),
),
as.Date("2021-03-01"),
"epiweek")}
}
\description{
Simple example forecaster that forecasts confirmed cases.
}
