% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{fieller}
\alias{fieller}
\title{Fieller's Confidence Interval}
\usage{
fieller(x, y, conf.level = 0.95, round = 2)
}
\arguments{
\item{x}{A numeric vector of the numerator}

\item{y}{A numeric vector of the denominator}

\item{conf.level}{confidence level of the vector}

\item{round}{number of decimal places to use}
}
\value{
A vector of the lower/upper limits of the confidence interval
}
\description{
Calculates confidence interval of a ratio of vectors under a normal distribution
}
