\name{FormulaParts-class}
\docType{class}
\alias{FormulaParts-class}

\title{Class "FormulaParts"}
\description{A class describing the parts of a formula.}
\section{Objects from the Class}{
    Objects can be created by calls of the form \code{new("FormulaParts", ...)}.
}
\section{Slots}{
  \describe{
    \item{\code{formula}:}{Object of class \code{"list"}.}
  }
}
\section{Methods}{
No methods defined with class "FormulaParts" in the signature.
}
\keyword{classes}
