% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{get_target_period}
\alias{get_target_period}
\title{Get the target period for a forecast date, incidence period and ahead}
\usage{
get_target_period(forecast_date, incidence_period, ahead)
}
\description{
Get the target period for a forecast date, incidence period and ahead
}
