% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{get_train_end_date}
\alias{get_train_end_date}
\title{Get last possible date of training data for given aheads.}
\usage{
get_train_end_date(lagged_df, ahead)
}
\arguments{
\item{lagged_df}{Data frame of lagged data as in \code{create_train_and_predict_matrices()}.}

\item{ahead}{Number of days ahead to forecast. Can be a single positive
integer or a vector of positive integers.}
}
\value{
Single date corresponding to the last possible date of the training
data.
}
\description{
Returns the last possible date of the training data for a given set of
aheads. If more than one ahead is given, the date return corresponds to the
last possible date corresponding to the smallest ahead value. This is because
the smallest ahead value will have the latest possible date.
}
