% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoff.R
\name{getContinuousPayoff}
\alias{getContinuousPayoff}
\title{Continuous discount function}
\usage{
getContinuousPayoff(discRate, t1, t2)
}
\arguments{
\item{discRate}{Discount rate}

\item{t1}{First time point}

\item{t2}{Second time point}
}
\value{
Discount factor
}
\description{
Continuous discounting function
}
