% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom-generics.R
\name{glance.bgw_mle}
\alias{glance.bgw_mle}
\title{Glance a \code{bgw_mle} object}
\usage{
\method{glance}{bgw_mle}(x, ...)
}
\arguments{
\item{x}{An object of class \code{bgw_mle}}

\item{...}{Additional arguments}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the components of the \code{bgw_mle} object
}
\description{
Glance a \code{bgw_mle} object
}
