% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{has_scores}
\alias{has_scores}
\title{Check if the model object contains the scores}
\usage{
has_scores(x)
}
\arguments{
\item{x}{A model object of class \code{bgw_mle}}
}
\value{
A boolean value equal to TRUE if the object contains the scores.
}
\description{
The scores matrix contains the gradient observations and are used to calcualte
the robust variance-covariance matrix. The function is only implemented and
used for objects of class \code{bgw_mle}. If called on other objects, the function
will throw an error.
}
