% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition.R
\name{lerp}
\alias{lerp}
\title{Linear Interpolation}
\usage{
lerp(x, x1, x2, y1, y2)
}
\arguments{
\item{x}{Input value to interpolate corresponding location on a curve.}

\item{x1}{Lower x bound for curve y~x.}

\item{x2}{Upper x bound for curve y~x.}

\item{y1}{Lower y value given x1 for curve y~x.}

\item{y2}{Upper y value given x2 for curve y~x.}
}
\value{
An interpolated value given two points on a curve.
}
\description{
Linear interpolation between two points on a curve. 
May be faster than base R function since only two points are used for interpolation.
}
