% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-generics.R
\name{logLik.bgw_mle}
\alias{logLik.bgw_mle}
\title{Log-likelihood of a \code{bgw_mle} object}
\usage{
\method{logLik}{bgw_mle}(object, ...)
}
\arguments{
\item{object}{An object of class \code{bgw_mle}}

\item{...}{Additional arguments}
}
\value{
An object of class \code{logLik} with attributes \code{df} (number of
parameters) and \code{nobs} (number of observations)
}
\description{
Retrieve the log likelihood value of the model. The value is returned as an
object of class \code{logLik} such that \code{\link[stats:AIC]{stats::AIC()}}, \code{\link[stats:BIC]{stats::BIC()}} and
likelihood ratio tests work out of the box.
}
