% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich-estimator.R
\name{meat}
\alias{meat}
\title{Meat for sandwiches}
\usage{
meat(x, adjust = FALSE, ...)
}
\arguments{
\item{x}{A model object of class \code{bgw_mle}}

\item{adjust}{A logical value indicating if the finite sample adjustment should
be applied to the meat matrix}

\item{...}{Additional arguments passed to methods}
}
\value{
A matrix
}
\description{
Meat for sandwiches
}
