% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Min}
\alias{Min}
\alias{Max}
\alias{Sum}
\alias{Mean}
\alias{Median}
\title{Wrappers to functions that remove \code{NA} values by default}
\usage{
Min(x)

Max(x)

Sum(x)

Mean(x)

Median(x)
}
\description{
Wrappers to functions that remove \code{NA} values by default
}
