% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltools-package.R
\docType{package}
\name{modeltools-package}
\alias{modeltools}
\alias{modeltools-package}
\title{modeltools: Tools for econometric modeling and visualization in R}
\description{
A collection of tools for econometric modeling and visualization in R with a particular focus on choice modeling using syntax consistent with the tidyverse.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://modeltools.edsandorf.me/}
}

}
\author{
\strong{Maintainer}: Erlend Dancke Sandorf \email{erlend.dancke.sandorf@nmbu.no} (\href{https://orcid.org/0000-0001-7521-4773}{ORCID})

Authors:
\itemize{
  \item Erlend Dancke Sandorf \email{erlend.dancke.sandorf@nmbu.no} (\href{https://orcid.org/0000-0001-7521-4773}{ORCID})
}

}
\keyword{internal}
