% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltools.R
\docType{package}
\name{modetools}
\alias{modetools}
\title{modeltools: Tools for Building COVID Forecasters and Related Models}
\description{
The `modeltools` package provides various tools for building COVID-19 models, 
mainly forecasting and hotspot prediction models. It is intended to be used
in conjuction with the `covidcast` and `evalcast` packages.
}
