% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-generics.R
\name{nobs.bgw_mle}
\alias{nobs.bgw_mle}
\title{Number of observations of a \code{bgw_mle} object}
\usage{
\method{nobs}{bgw_mle}(object, ...)
}
\arguments{
\item{object}{An object of class \code{bgw_mle}}

\item{...}{Additional arguments}
}
\value{
A single integer
}
\description{
Number of observations of a \code{bgw_mle} object
}
