% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{pct_change}
\alias{pct_change}
\title{Compute percentage change of values in `covidcast_signal` data frame}
\usage{
pct_change(x, n = 14, new_col = "pct_change")
}
\arguments{
\item{x}{The `covidcast_signal` data frame under consideration.}

\item{n}{Size of the local window (in days) to use (bumped up to the nearest
even integer, if needed). For example, if `n = 10`, then to compute the
percentage change on November 10, we use 100 * (B - A) / A, where A is the
sum of the values between November 6 and November 10, and A is the sum of
the values between November 1 and November 5. Default is 14.}

\item{new_col}{String indicating the name of the new column that will contain
the percentage change values. Default is "pct_change"; note that we can set
`new_col = "value"` to overwrite the existing "value" column.}
}
\value{
A data frame given by appending a new column to `x` named according
  to the `new_col` argument, containing the percentage change values.
}
\description{
Computes the percentage change of the values in a `covidcast_signal` data
frame. (When multiple issue dates are present, only the latest issue is
considered.)  See the [percentage change vignette]() for examples.
}
