% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poe.R
\name{print.poe_test}
\alias{print.poe_test}
\title{Generic print method for \code{poe_test()}}
\usage{
\method{print}{poe_test}(x, ...)
}
\arguments{
\item{x}{An object of class \code{poe_test}}

\item{...}{Additional arguments}
}
\description{
Generic print method for \code{poe_test()}
}
