% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{ProbToOdds}
\alias{ProbToOdds}
\title{Convert probability to odds}
\usage{
ProbToOdds(p)
}
\arguments{
\item{p}{A vector of probabilities}
}
\value{
A vector of odds.
}
\description{
Converts a vector of probabilities to odds.
}
\details{
If any probabilities are less than 0 or greater than 1, the function returns a \emph{NULL}.
}
