% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{ProbToRate}
\alias{ProbToRate}
\title{Convert a probability to a rate}
\usage{
ProbToRate(p, t = 1)
}
\arguments{
\item{p}{A vector of probabilities.}

\item{t}{The time frame of the probability.}
}
\value{
A vector of rates in the unit of time \emph{t}.
}
\description{
Converts a vector of probabilities to a vector of rates.
}
\details{
If any probabilities are less than 0 or greater than 1, the function returns \emph{NULL}.
}
