% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantreg.R
\name{quantreg_forecaster}
\alias{quantreg_forecaster}
\title{Simple quantile autoregressive forecaster template based on \code{quantgen}
(unpenalized)}
\usage{
quantreg_forecaster(
  df,
  forecast_date,
  ahead,
  tau = modeltools::covidhub_probs,
  n = 28,
  lags = 0,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{df}{Data frame of signal values to use for forecasting, of the format
that is returned by \code{\link[covidcast:covidcast_signals]{covidcast::covidcast_signals()}}.}

\item{forecast_date}{Date object or string of the form "YYYY-MM-DD",
indicating the date on which forecasts will be made. For example, if
\code{forecast_date = "2020-05-11"} and \code{ahead = 3}, then, forecasts
would be made for "2020-05-14".}

\item{ahead}{Vector of ahead values, indicating how many days ahead
to forecast. E.g. \code{ahead = 1} means the day after forecast date.}

\item{tau}{Vector of quantile levels for the probabilistic forecast. If not
specified, defaults to the levels required by the COVID Forecast Hub.}

\item{n}{Size of the local training window (in days) to use. For example,
if \code{n = 14}, then to make a 1-day-ahead forecast on December 15, we
train on data from December 1 to December 14.}

\item{lags}{Vector of lag values to use as features in the autoregressive
model. For example, setting \code{lags = c(0, 7, 14)} means we use the current
value of each signal, as well as the values 7 and 14 days ago, as the
features. Note that \code{lags} can also be a list of vectors of lag
values, this list having the same length as \code{df},
in order to apply a different set of shifts to each signal.
Default is 0, which means no additional lags (only current values) for each
signal.}

\item{verbose}{Should progress be printed out to the console? Default is
\code{FALSE}.}

\item{...}{Additional arguments.}
}
\value{
Data frame with columns \code{ahead}, \code{geo_value}, \code{quantile}, and
\code{value}. The \code{quantile} column gives the probabilities associated with
quantile forecasts for that location and ahead.
}
\description{
A simple quantile autoregressive forecaster template based on \code{quantgen}, to
be used with \code{evalcast}, via \code{\link[evalcast:get_predictions]{evalcast::get_predictions()}}. This is a
simplified, unpenalized version of \code{\link[=quantgen_forecaster]{quantgen_forecaster()}}
that is meant for illustration purposes only. Only works with
\code{incidence_period = "day"}.
}
