% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{RateToProb}
\alias{RateToProb}
\title{Convert a rate to a probability}
\usage{
RateToProb(r, t = 1)
}
\arguments{
\item{r}{A vector of rates}

\item{t}{The time frame of the desired probability}
}
\value{
A vector of probabilities in the unit of time \emph{t}.
}
\description{
Converts a vector of rates to a vector of probabilities.
}
\details{
If any rates are less than 0, the function returns \emph{NULL}.
}
