% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{repeat_rows}
\alias{repeat_rows}
\title{Repeat rows}
\usage{
repeat_rows(x, n)
}
\arguments{
\item{x}{A \link[base:matrix]{matrix}, \link[base:data.frame]{data.frame} or
\link[tibble:tibble-package]{tibble}}

\item{n}{The number of times to repeat each row}
}
\value{
A \link[base:matrix]{matrix}, \link[base:data.frame]{data.frame} or
\link[tibble:tibble-package]{tibble} depending on the type of the input
}
\description{
This function repeats the rows of a \link[base:matrix]{matrix},
\link[base:data.frame]{data.frame} or \link[tibble:tibble-package]{tibble} by a given
number of times.
}
\details{
In case of a data.frame, new row names are preserved.
}
\examples{
repeat_rows(matrix(1:4, nrow = 2), 2)
repeat_rows(tibble::tibble(a = 1:2, b = 3:4), 2)
repeat_rows(data.frame(a = 1:2, b = 3:4), 2)

}
