% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCR_utils.R
\name{ROCR_get_AUC}
\alias{ROCR_get_AUC}
\title{Retrieve AUC from the performance function}
\usage{
ROCR_get_AUC(pred)
}
\arguments{
\item{pred}{a ROCR prediction object
return numeric AUC value}
}
\description{
Retrieve AUC from the performance function
}

