% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition.R
\name{roundToMult}
\alias{roundToMult}
\title{Round to Specified Multiple}
\usage{
roundToMult(v, m)
}
\arguments{
\item{v}{A vector of values to be rounded.}

\item{m}{A multiple to round values to.}
}
\value{
A vector of values rounded to the nearest multiple, \emph{m}.
}
\description{
Round a vector to the nearest user-specified multiple.
}
