% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randvar.R
\name{rweireg}
\alias{rweireg}
\title{Random Weibull}
\usage{
rweireg(n, b, x, s)
}
\arguments{
\item{n}{The number of variates to sample.}

\item{b}{A vector of weibull regression coefficients.}

\item{x}{A vector of covariate values.}

\item{s}{Shape parameter.}
}
\value{
A vector of times to event.
}
\description{
Retrieves random variate from weibull model.
}
