% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich-estimator.R
\name{sandwich-methods}
\alias{sandwich-methods}
\alias{estfun.bgw_mle}
\alias{bread.bgw_mle}
\title{Sandwich estimator methods for \code{bgw_mle} objects}
\usage{
\method{estfun}{bgw_mle}(x, ...)

\method{bread}{bgw_mle}(x, ...)
}
\arguments{
\item{x}{A model object of class \code{bgw_mle}}

\item{...}{Additional arguments passed to methods}
}
\value{
\code{estfun()} returns the scores matrix. \code{bread()} returns the
variance-covariance matrix multiplied by the number of observations.
}
\description{
Methods for the \code{\link[sandwich:estfun]{sandwich::estfun()}} and \code{\link[sandwich:bread]{sandwich::bread()}} generics.
These make all variance-covariance estimators in the \code{sandwich} package
available for objects of class \code{bgw_mle}, e.g., \code{\link[sandwich:sandwich]{sandwich::sandwich()}} for
the robust variance-covariance matrix and \code{\link[sandwich:vcovCL]{sandwich::vcovCL()}} for
clustering at the individual level. The scores must first be added to the
model object using \code{\link[=add_scores]{add_scores()}}.
}
\details{
Note that the default Hessian approximation in \code{\link[bgw:bgw_mle]{bgw::bgw_mle()}} is BHHH, in
which case the robust variance-covariance matrix equals \code{vcov(x)}. Estimate
the model with \code{bgw_settings = list(vcHessianMethod = "finiteDifferences")}
to obtain a robust variance-covariance matrix that differs from \code{vcov(x)}.
}
\examples{
\dontrun{
  model <- add_scores(model, log_lik)
  sandwich(model)
  sandwich(model, adjust = TRUE)
  vcovCL(model, cluster = db$id)
}

}
