% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich-estimator.R
\name{sandwich}
\alias{sandwich}
\title{Sandwich estimator}
\usage{
sandwich(x, ...)
}
\arguments{
\item{x}{A fitted model object}

\item{...}{Additional arguments passed to methods}
}
\value{
A robust variance-covariance matrix
}
\description{
An implementation of the sandwich estimator to use when calculating the robust
variance-covariance matrix for the sandwich estimator when using the BGW algorithm
}
