% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich-estimator.R
\name{scores}
\alias{scores}
\title{Extract the empirical estimation function}
\usage{
scores(x, ...)
}
\arguments{
\item{x}{A model object of class \code{bgw_mle}}

\item{...}{Additional arguments passed to methods}
}
\value{
A matrix with one row per observation and one column per parameter
}
\description{
This function extracts the scores matrix from a model object of class
\code{bgw_mle}. The scores matrix contains the gradient observations and is used
to calculate the robust variance-covariance matrix.
}
