% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{slide_by_geo}
\alias{slide_by_geo}
\title{Slide a function over values in `covidcast_signal` data frame, grouped by
`geo_value`}
\usage{
slide_by_geo(df, slide_fun, n = 14, new_col = "slide_value", params = list())
}
\arguments{
\item{df}{The `covidcast_signal` data frame under consideration.}

\item{slide_fun}{Function to slide over the values in `df`, grouped by
`geo_value`. To "slide" means to apply the function over a trailing window
of `n` days (at any time point, we run the function over the last `n` days
of data).}

\item{n}{Size of the local window (in days) to use. For example, if `n = 5`,
then to estimate the derivative on November 5, we train the given method on
data in between November 1 and November 5. Default is 14.}

\item{new_col}{String indicating the name of the new column that will contain
the derivative values. Default is "slide_value"; note that setting `new_col
= "value"` will overwrite the existing "value" column.}

\item{params}{List of additional arguments to pass to the function specified
via `slide_fun`.}
}
\value{
A data frame given by appending a new column to `df` named according
  to the `new_col` argument, containing the function values.
}
\description{
Slides a given function over the values in a  `covidcast_signal` data frame,
grouped by `geo_value`. (When multiple issue dates are present, only the
latest issue is considered.)  See the [slide vignette]() for examples.
}
