% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{stars}
\alias{stars}
\title{Create stars}
\usage{
stars(p)
}
\arguments{
\item{p}{A double or numeric vector of p-values.}
}
\value{
A character vector of stars.
}
\description{
Creates stars based on the p-value of a test.
}
\examples{
stars(0.05)

}
