\name{StatModelCapabilities-class}
\docType{class}
\alias{StatModelCapabilities-class}

\title{Class "StatModelCapabilities" }
\description{ A class describing capabilities of a statistical model. }
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("StatModelCapabilities", ...)}.
}
\section{Slots}{
  \describe{
    \item{\code{weights}:}{Object of class \code{"logical"}}
    \item{\code{subset}:}{Object of class \code{"logical"}}
  }
}
\section{Methods}{
No methods defined with class "StatModelCapabilities" in the signature.
}
\keyword{classes}
