% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{suggest_parallel}
\alias{suggest_parallel}
\title{Function to suggest parallel settings}
\usage{
suggest_parallel()
}
\value{
A list with suggested settings
}
\description{
Suggests the number of cores and cluster type to use for parallel
processing based on the operating system and the number of cores available.
}
