% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom-generics.R
\name{tidy.bgw_mle}
\alias{tidy.bgw_mle}
\title{Tidy a \code{bgw_mle} object}
\usage{
\method{tidy}{bgw_mle}(x, vcov = stats::vcov(x), ...)
}
\arguments{
\item{x}{An object of class \code{bgw_mle}}

\item{vcov}{A variance-covariance matrix used to calculate the standard
errors. Defaults to \code{vcov(x)}. Use, e.g., \code{sandwich(x)} for robust standard
errors.}

\item{...}{Additional arguments}
}
\value{
A tidy \code{\link[tibble:tibble]{tibble::tibble()}} with the components of the \code{bgw_mle} object
}
\description{
Standard errors are calculated from the supplied variance-covariance matrix
and p-values are based on the asymptotic normal distribution.
}
