% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-generics.R
\name{vcov.bgw_mle}
\alias{vcov.bgw_mle}
\title{Variance-covariance matrix of a \code{bgw_mle} object}
\usage{
\method{vcov}{bgw_mle}(object, ...)
}
\arguments{
\item{object}{An object of class \code{bgw_mle}}

\item{...}{Additional arguments}
}
\value{
A numeric matrix with the variances and covariances of the estimated
parameters
}
\description{
Variance-covariance matrix of a \code{bgw_mle} object
}
