% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-object-modifiers.R
\name{add_scores}
\alias{add_scores}
\title{Calculates and adds the scores to a fitted model object}
\usage{
add_scores(object, func, x, ...)
}
\arguments{
\item{object}{A fitted model object of class}

\item{func}{A function with real (vector) results. This is typically the
log-likelihood function. The function must return the function values at the
observation level.}

\item{x}{A real or real vector argument to func, indicating the point at
which the gradient is to be calculated.}

\item{...}{Additional arguments passed to \code{\link[numDeriv:jacobian]{numDeriv::jacobian()}}.}
}
\value{
A fitted model object with the scores added to the
object.
}
\description{
Calculates and adds the scores to a fitted model object of.
The scores are the gradient observations/first derivatives of the log-likelihood
function. The function is a wrapper around \code{\link[numDeriv:jacobian]{numDeriv::jacobian()}}.
}
