% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{converged}
\alias{converged}
\title{Check if model has converged}
\usage{
converged(x, ...)
}
\arguments{
\item{x}{A model object}

\item{...}{Additional arguments}
}
\value{
A logical value indicating if the model has converged
}
\description{
Using the return code from the optimizer, check if the code indicates that
the optimization converged to a local optimum.
}
\details{
Objects of class \code{bgw_mle} will have the following codes:
0 - Initial f(x) cannot be computed
4 - Relative function convergence
5 - X- and relative function convergence
7 - Singular convergence
8 - False convergence
9 - Function evaluation limit
}
