% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv.R
\name{estimate_deriv}
\alias{estimate_deriv}
\title{Estimate derivatives of values in \code{covidcast_signal} data frame}
\usage{
estimate_deriv(
  x,
  method = c("lin", "ss", "tf"),
  n = 14,
  col_name = "deriv",
  keep_obj = FALSE,
  deriv = 1,
  ...
)
}
\arguments{
\item{x}{The \code{covidcast_signal} data frame under consideration.}

\item{method}{One of "lin", "ss", or "tf" indicating the method to use for
the derivative calculation. To estimate the derivative at any time point,
we run the given method on the last \code{n} days of data, and use the
corresponding predicted derivative (that is, the derivative of the
underlying estimated function, linear or spline) at the current time
point. Default is "lin". See details below.}

\item{n}{Size of the local window (in days) to use. For example, if \code{n = 5},
then to estimate the derivative on November 5, we train the given method on
data in between November 1 and November 5. Default is 14.}

\item{col_name}{String indicating the name of the new column that will
contain the derivative values. Default is "deriv"; note that setting
\code{col_name = "value"} will overwrite the existing "value" column.}

\item{keep_obj}{Should the fitted object (from linear regression, smoothing
spline, or trend filtering) be kept as a separate column? If \code{TRUE}, then
this column name is given by  appending "_obj" to \code{col_name}. Default is
\code{FALSE}.}

\item{deriv}{Order of derivative to estimate. Only orders 1 or 2 are allowed,
with the default being 1. (In some cases, a second-order derivative will
return a trivial result: for example: when \code{method = "lin"}, this will
always be zero.)}

\item{...}{Additional arguments to pass to the function that estimates
derivatives. See details below.}
}
\value{
A data frame given by appending a new column to \code{x} named according
to the \code{col_name} argument, containing the estimated derivative values.
}
\description{
Estimates derivatives of the values in a \code{covidcast_signal} data frame, using
a local (in time) linear regression or smoothing spline. (When multiple issue
dates are present, only the latest issue is considered.)  See the
\href{https://cmu-delphi.github.io/covidcast/modeltoolsR/articles/estimate-deriv.html}{estimating derivatives vignette}
for examples.
}
\details{
Derivatives are estimated using:

\itemize{
\item Linear regression, when \code{method = "lin"}, via \code{stats::lsfit()}.
\item Cubic smoothing spline, when \code{method = "ss"}, via
\code{stats::smooth.spline()}.
\item Polynomial trend filtering, when \code{method = "tf"}, via
\code{genlasso::trendfilter()}.
}

The second and third cases base the derivative calculation on a nonparametric
fit and should typically be used with a larger window \code{n}. The third case
(trend filtering) is more locally adaptive than the second (smoothing
spline) and can work better when there are sharp changes in the smoothness
of the underlying values.

In the first and second cases (linear regression and smoothing spline), the
additional arguments in \code{...} are directly passed to the underlying
estimation function (\code{stats::lsfit()} and \code{stats::smooth.spline()}).

The third case (trend filtering) works a little differently. Here, a custom
set of arguments is allowed (and are internally distributed as appropriate
to \code{genlasso::trendfilter()}, \code{genlasso::cv.trendfilter()}, and
\code{genlasso::coef.genlasso()}):

\describe{
\item{\code{ord}}{Order of piecewise polynomial for the trend filtering fit,
default is 2.}
\item{\code{maxsteps}}{Maximum number of steps to take in the solution path before
terminating, default is 100.}
\item{\code{cv}}{Boolean indicating whether cross-validation should be used to
choose an effective degrees of freedom for the fit, default is \code{FALSE}.}
\item{\code{k}}{Number of folds if cross-validation is to be used. Default is 5.}
\item{\code{df}}{Desired effective degrees of freedom for the trend filtering
fit. If \code{cv = FALSE}, then \code{df} must be an integer; if \code{cv = TRUE}, then
\code{df} should be one of "min" or "1se" indicating the selection rule to use
based on the cross-validation error curve (minimum or 1-standard-error
rule, respectively). Default is 8 when \code{cv = FALSE}, and "1se" when \code{cv = TRUE}.}
}
}
\examples{
\dontrun{
df <- covidcast::covidcast_signal("fb-survey", "smoothed_cli", 
    start_day = "2021-01-01", 
    end_day = "2021-02-28",
    geo_type = "state",
    geo_values = c("ca", "fl"))

# estimate derivative using linear regression and n = 10 days
estimate_deriv(df, method = "lin", n = 10)

# keep the linear regression fits
estimate_deriv(df, method = "lin", n = 10, keep_obj = TRUE)

# estimate derivative using smoothing spline with 8 degrees of freedom
estimate_deriv(df, method = "ss", n = 28, df = 8)

# estimate derivative using trend filtering with 8 degrees of freedom
estimate_deriv(df, method = "tf", n = 28, df = 8)
}

}
