% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{has_scores}
\alias{has_scores}
\title{Check if the model object contains the scores}
\usage{
has_scores(x)
}
\arguments{
\item{x}{A model object}
}
\value{
A boolean value equal to TRUE if the object contains the scores.
}
\description{
The scores matrix contains the gradient observations and are used to calculate
the robust variance-covariance matrix. The function checks whether the scores
matrix is present in the model object. The function is primarily used to
control the flow in other functions.
}
