% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_bgw}
\alias{is_bgw}
\title{Check if the model object is of class \code{bwg_mle}}
\usage{
is_bgw(x)
}
\arguments{
\item{x}{A model object}
}
\value{
A Boolean value equal to TRUE if the object is of class \code{bgw_mle}.
}
\description{
A simple function checking the class of the object. The function is primarily
used to control flow in other functions.
}
