% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_bgw}
\alias{is_bgw}
\title{Check if the model object is of class \code{bwg_mle}}
\usage{
is_bgw(x)
}
\arguments{
\item{x}{A model object of class \code{bgw_mle}}
}
\value{
A boolean value equal to TRUE if the object is of class \code{bgw_mle}.
}
\description{
Check if the model object is of class \code{bwg_mle}
}
