% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_packages}
\alias{load_packages}
\title{Load packages}
\usage{
load_packages(pkgs, install_missing = TRUE)
}
\arguments{
\item{pkgs}{A character vector of package names}

\item{install_missing}{A boolean indicating whether to install missing
packages. Default is TRUE}
}
\value{
The function does not return anything
}
\description{
A wrapper for \code{library()} that can load multiple packages at once. If the
package is not installed, the function will install it first.
}
\examples{
\dontrun{
   load_packages(c("dplyr", "ggplot2"))
 }
 
}
