% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-generics.R
\name{logLik.bgw_mle}
\alias{logLik.bgw_mle}
\title{Log-likelihood of a \code{bgw_mle} object}
\usage{
\method{logLik}{bgw_mle}(object, ...)
}
\arguments{
\item{object}{An object of class \code{bgw_mle}}

\item{...}{Additional arguments}
}
\value{
A single double
}
\description{
Retrieve the log likelihood value of the model.
}
