% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{OddsToProb}
\alias{OddsToProb}
\title{Convert odds to probability}
\usage{
OddsToProb(o)
}
\arguments{
\item{o}{A vector of odds.}
}
\value{
A vector of probabilities.
}
\description{
Converts a vector of odds to probabilities.
}
\details{
If any odds are less than 0, the function returns \emph{NULL}.
}
