% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{pct_change}
\alias{pct_change}
\title{Compute percentage change of values in \code{covidcast_signal} data frame}
\usage{
pct_change(x, n = 14, col_name = "pct_change")
}
\arguments{
\item{x}{The \code{covidcast_signal} data frame under consideration.}

\item{n}{Size of the local window (in days) to use (bumped up to the nearest
even integer, if needed). For example, if \code{n = 10}, then to compute the
percentage change on November 10, we use 100 * (B - A) / A, where A is the
sum of the values between November 6 and November 10, and A is the sum of
the values between November 1 and November 5. Default is 14.}

\item{col_name}{String indicating the name of the new column that will
contain the percentage change values. Default is "pct_change"; note that we
can set \code{col_name = "value"} to overwrite the existing "value" column.}
}
\value{
A data frame given by appending a new column to \code{x} named according
to the \code{col_name} argument, containing the percentage change values.
}
\description{
Computes the percentage change of the values in a \code{covidcast_signal} data
frame. (When multiple issue dates are present, only the latest issue is
considered.)  See the \href{https://cmu-delphi.github.io/covidcast/modeltoolsR/articles/pct-change.html}{percentage change vignette}
for examples.
}
\examples{
\dontrun{
df <- covidcast::covidcast_signal("fb-survey", "smoothed_cli", 
    start_day = "2021-01-01", 
    end_day = "2021-01-31",
    geo_type = "state")

# percentage change between back-to-back weeks (default, as n = 14)
pct_change(df)

# percentage change between back-to-back days
pct_change(df, n = 2)
}

}
