% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poe.R
\name{poe_test}
\alias{poe_test}
\title{Test for the difference in independent empirical distributions}
\usage{
poe_test(x, y)
}
\arguments{
\item{x, y}{Two numeric vectors representing the independent empriical
distributions to be compared}
}
\value{
A list with the following components:
\itemize{
\item \code{method} The name of the test
\item \code{statistic} The test statistic
\item \code{means} The means of the two input vectors
}
}
\description{
The function calculates the complete combinatorial of the supplied vectors
using a loop implementation by taking the difference between every
combination of the elements in the vectors. The input vectors must
be numeric, but can be of different lengths.
}
\examples{
x <- qnorm(runif(100), mean = -0.5, sd = 1)
y <- qnorm(runif(100), mean = 1.5, sd = 2)
poe_test(x, y)
 
}
\references{
Poe G. L., Giraud, K. L. & Loomis, J. B., 2005, Computational methods for
measuring the difference of empirical distributions, American Journal of
Agricultural Economics, 87:353-365
}
