% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{prep_for_gt}
\alias{prep_for_gt}
\title{Prepare model output for \code{gt}}
\usage{
prep_for_gt(x, ...)
}
\arguments{
\item{x}{A model object}

\item{...}{Additional arguments passed to \code{\link[generics:tidy]{generics::tidy()}} and
\code{\link[generics:glance]{generics::glance()}}, e.g., \code{vcov = sandwich(x)} for robust standard errors}
}
\value{
A tibble with estimated coefficients, standard errors, and stars.
}
\description{
Takes a model object as an input and prepares a table with estimated
coefficients, standard errors, and stars indicating significance. Names in
\code{term} will be converted to lower case to ensure seamless use of left join.
}
\details{
The table can quickly chained to other tables using left_join().
}
