% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom-generics.R, R/sandwich-estimator.R,
%   R/stats-generics.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tidy}
\alias{glance}
\alias{estfun}
\alias{bread}
\alias{meat}
\alias{sandwich}
\alias{vcovCL}
\alias{logLik}
\alias{coef}
\alias{vcov}
\alias{nobs}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics:glance]{glance()}}, \code{\link[generics:tidy]{tidy()}}}

  \item{sandwich}{\code{\link[sandwich:bread]{bread()}}, \code{\link[sandwich:estfun]{estfun()}}, \code{\link[sandwich:meat]{meat()}}, \code{\link[sandwich:sandwich]{sandwich()}}, \code{\link[sandwich:vcovCL]{vcovCL()}}}

  \item{stats}{\code{\link[stats:coef]{coef()}}, \code{\link[stats:logLik]{logLik()}}, \code{\link[stats:nobs]{nobs()}}, \code{\link[stats:vcov]{vcov()}}}
}}

