% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starting-values.R
\name{search_starting_values}
\alias{search_starting_values}
\title{Function to search for starting values}
\usage{
search_starting_values(
  log_lik,
  starting_values,
  N = 10000,
  return = 10,
  adjustment_multiplier = 1
)
}
\arguments{
\item{log_lik}{A log likelihood function}

\item{starting_values}{A vector of starting values. Note: Named vectors are ignored. The startign values must be locally bound in the log-likelihood function to work correctly.}

\item{N}{The number of starting value candiates to use. Default is 10,000}

\item{return}{Number of vectors to return}

\item{adjustment_multiplier}{An adjustment multiplier for the random uniform adjustment matrix. The default value is 1.}
}
\value{
A list of \code{return} number of parametrs
}
\description{
The function takes a vector of starting values. These will then be adjusted by adding a random unform
value between -1 and 1 multiplied by an adjustment multiplier. The function then evaluates the
log-likelihood for each set of starting values and returns the specified number of best fitting
starting values.
}
