% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{slide_by_geo}
\alias{slide_by_geo}
\title{Slide a function over values in \code{covidcast_signal} data frame, grouped by
\code{geo_value}}
\usage{
slide_by_geo(
  x,
  slide_fun,
  n = 14,
  shift = 0,
  col_name = "slide_value",
  col_type = c("dbl", "int", "lgl", "chr", "list"),
  ...
)
}
\arguments{
\item{x}{The \code{covidcast_signal} data frame under consideration.}

\item{slide_fun}{Function or formula to slide over the values in \code{x}, grouped
by \code{geo_value}. To "slide" means to apply the function or formula over a
right-aligned window of \code{n} days of data. If a function, \code{slide_fun} must take
\code{x}, a data frame the same column names as the original data frame;
followed by any number of named additional arguments; and ending with
\code{...}, to capture general additional arguments. If a formula, \code{slide_fun}
can operate directly on \code{.x$value}, \code{.x$time_value}, etc., as in \code{~ mean(.x$value)} to compute a trailing mean over the last \code{n} days of data.}

\item{n}{Size of the local window (in days) to use.  Default is 14.}

\item{shift}{Number of days forward from the \code{time_value} to shift the window.  Because the
window is right-aligned, if \code{n = 5} and \code{shift = 0}, then for a \code{time_value} of November 5,
the function will be applied to values from November 1 to November 5.  A positive value shifts
the window forwards in time, while negative values shift the window backwards in time.  Ex:
for a \code{time_value} of November 5 \code{shift = 3} captures November 4 through November 8;
\code{shift = -3} captures October 29 through November 2.  Default is not to shift.}

\item{col_name}{String indicating the name of the new column that will
contain the derivative values. Default is "slide_value"; note that setting
\code{col_name = "value"} will overwrite the existing "value" column.}

\item{col_type}{One of "dbl", "int", "lgl", "chr", or "list", indicating the
data type (as tibble abbreviation) for the new column. Default is "dbl".}

\item{...}{Additional arguments to pass to the function or formula specified
via \code{slide_fun}.}
}
\value{
A data frame given by appending a new column to \code{x} named according
to the \code{col_name} argument, containing the function values.
}
\description{
Slides a given function over the values in a \code{covidcast_signal} data frame,
grouped by \code{geo_value}. (When multiple issue dates are present, only the
latest issue is considered.) See the \href{https://cmu-delphi.github.io/covidcast/modeltoolsR/articles/modeltools.html}{getting started guide}
for examples.
}
\examples{
\dontrun{
df <- covidcast::covidcast_signal("fb-survey", "smoothed_cli", 
    start_day = "2021-01-01", 
    end_day = "2021-01-31",
    geo_type = "state")
    
# two equivalent ways to compute 7-day trailing averages
slide_by_geo(df, slide_fun = ~ Mean(.x$value), n = 7)
slide_by_geo(df, slide_fun = function(x, ...) Mean(x$value) , n = 7)
}

}
