% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom-generics.R
\name{tidy.bgw_mle}
\alias{tidy.bgw_mle}
\title{Tidy a `bgw_mle´ object}
\usage{
\method{tidy}{bgw_mle}(x, ...)
}
\arguments{
\item{x}{An object of class \code{bgw_mle}}

\item{...}{Additional arguments}
}
\value{
A tidy \code{\link[tibble:tibble]{tibble::tibble()}} with the components of the \code{bgw_mle} object
}
\description{
Tidy a `bgw_mle´ object
}
